+++ src/entities/user/api/actions.ts
@@ -5,8 +5,7 @@ import { revalidatePath } from "next/cache";
 import { TwinsAPI } from "@/shared/api";
 import { isPopulatedArray } from "@/shared/libs";
 
-import { hydrateDomainUserFromMap } from "../libs/helpers";
-import { LOGIN_FORM_SCHEMA } from "../libs/schemas";
+import { LOGIN_FORM_SCHEMA, hydrateDomainUserFromMap } from "../server";
 
 async function login(authToken: string, domainId: string) {
   const { data, error } = await TwinsAPI.POST(
